requirements.txt
gradio
transformers
torch
pillow
nltk
scikit-learn
openai
faiss-cpu
speechrecognition
gtts
pyttsx3
python-dotenv
